fix: log job when storage file missing for restoration
